Module snippets::snippets

DO NOT EDIT! Code generated by docs/snippets/build.rs.

Modules

annotation_context_connections 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/annotation_context_connections.rs. Log annotation context with connections between keypoints.

annotation_context_rects 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/annotation_context_rects.rs. Log rectangles with different colors and labels using annotation context

annotation_context_segmentation 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/annotation_context_segmentation.rs. Log a segmentation image with annotations.

arrow2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/arrow2d_simple.rs. Log a batch of 2D arrows.

arrow3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/arrow3d_simple.rs. Log a batch of 3D arrows.

asset3d_out_of_tree 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/asset3d_out_of_tree.rs. Log a simple 3D asset with an out-of-tree transform which will not affect its children.

asset3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/asset3d_simple.rs. Log a simple 3D asset.

bar_chart 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/bar_chart.rs. Create and log a bar chart

box2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/box2d_simple.rs. Log some very simple 2D boxes.

box3d_batch 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/box3d_batch.rs. Log a batch of oriented bounding boxes.

box3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/box3d_simple.rs. Log a single 3D box.

clear_recursive 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/clear_recursive.rs. Log and then clear data recursively.

clear_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/clear_simple.rs. Log and then clear data.

custom_data 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/tutorials/custom_data.rs. Demonstrates how to implement custom archetypes and components, and extend existing ones.

depth_image_3d 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/depth_image_3d.rs. Create and log a depth image.

depth_image_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/depth_image_simple.rs. Create and log a depth image.

disconnected_space 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/disconnected_space.rs. Disconnect two spaces.

entity_path 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/entity_path.rs. Example of different ways of constructing an entity path.

image_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/image_simple.rs. Create and log an image

line_segments2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_segments2d_simple.rs. Log a couple 2D line segments using 2D line strips.

line_segments3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_segments3d_simple.rs. Log a simple set of line segments.

line_strip2d_batch 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strip2d_batch.rs. Log a batch of 2D line strips.

line_strip2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strip2d_simple.rs. Log a simple line strip.

line_strip3d_batch 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strip3d_batch.rs. Log a batch of 2D line strips.

line_strip3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strip3d_simple.rs. Log a simple line strip.

manual_indicator 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/manual_indicator.rs. Shows how to manually associate one or more indicator components with arbitrary data.

mesh3d_indexed 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/mesh3d_indexed.rs. Log a simple colored triangle with indexed drawing.

mesh3d_partial_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/mesh3d_partial_updates.rs. Log a simple colored triangle, then update its vertices’ positions each frame.

mesh3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/mesh3d_simple.rs. Log a simple colored triangle.

pinhole_perspective 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/pinhole_perspective.rs. Logs a point cloud and a perspective camera looking at it.

pinhole_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/pinhole_simple.rs. Log a pinhole and a random image.

point2d_random 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/point2d_random.rs. Log some random points with color and radii.

point2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/point2d_simple.rs. Log some very simple points.

point3d_random 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/point3d_random.rs. Log some random points with color and radii.

point3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/point3d_simple.rs. Log some very simple points.

quick_start_connect 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/quick_start/quick_start_connect.rs. Connect to the viewer and log some data.

quick_start_spawn 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/quick_start/quick_start_spawn.rs. Spawn a viewer and log some data.

scalar_multiple_plots 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/scalar_multiple_plots.rs. Log a scalar over time.

scalar_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/scalar_simple.rs. Log a scalar over time.

segmentation_image_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/segmentation_image_simple.rs. Create and log a segmentation image.

series_line_style 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/series_line_style.rs. Log a scalar over time.

series_point_style 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/series_point_style.rs. Log a scalar over time.

tensor_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/tensor_simple.rs. Create and log a tensor.

text_document 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/text_document.rs. Log a TextDocument

text_log 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/text_log.rs. Log a TextLog

text_log_integration 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/text_log_integration.rs. Shows integration of Rerun’s TextLog with the native logging interface.

transform3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/transform3d_simple.rs. Log different transforms between three arrows.

view_coordinates_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/view_coordinates_simple.rs. Change the view coordinates for the scene.

Functions

run