snippets
Module snippets

DO NOT EDIT! Code generated by docs/snippets/build.rs.

Modules

annotation_context 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/tutorials/annotation_context.rs.

annotation_context_connections 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/annotation_context_connections.rs. Log annotation context with connections between keypoints.

annotation_context_rects 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/annotation_context_rects.rs. Log rectangles with different colors and labels using annotation context

annotation_context_segmentation 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/annotation_context_segmentation.rs. Log a segmentation image with annotations.

any_batch_value_column_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/howto/any_batch_value_column_updates.rs. Use send_column to send an entire column of custom data to Rerun.

any_values 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/tutorials/any_values.rs. Log arbitrary data.

any_values_column_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/howto/any_values_column_updates.rs. Update custom user-defined values over time, in a single operation.

any_values_row_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/howto/any_values_row_updates.rs. Update custom user-defined values over time.

arrows2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/arrows2d_simple.rs. Log a batch of 2D arrows.

arrows3d_column_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/arrows3d_column_updates.rs. Update a set of vectors over time, in a single operation.

arrows3d_row_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/arrows3d_row_updates.rs. Update a set of vectors over time.

arrows3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/arrows3d_simple.rs. Log a batch of 3D arrows.

asset3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/asset3d_simple.rs. Log a simple 3D asset.

bar_chart 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/bar_chart.rs. Create and log a bar chart

boxes2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/boxes2d_simple.rs. Log some very simple 2D boxes.

boxes3d_batch 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/boxes3d_batch.rs. Log a batch of oriented bounding boxes.

boxes3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/boxes3d_simple.rs. Log a single 3D box.

capsules3d_batch 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/capsules3d_batch.rs. Log a batch of capsules.

clear_recursive 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/clear_recursive.rs. Log and then clear data recursively.

clear_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/clear_simple.rs. Log and then clear data.

custom_data 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/tutorials/custom_data.rs. Demonstrates how to implement custom archetypes and components, and extend existing ones.

dataframe_query 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/reference/dataframe_query.rs. Query and display the first 10 rows of a recording.

dataframe_view_query_external 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/reference/dataframe_view_query_external.rs. Query and display the first 10 rows of a recording in a dataframe view.

depth_image_3d 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/depth_image_3d.rs. Create and log a depth image.

depth_image_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/depth_image_simple.rs. Create and log a depth image.

descr_builtin_archetype 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/descriptors/descr_builtin_archetype.rs.

descr_builtin_component 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/descriptors/descr_builtin_component.rs.

descr_custom_archetype 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/descriptors/descr_custom_archetype.rs.

descr_custom_component 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/descriptors/descr_custom_component.rs.

different_data_per_timeline 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/concepts/different_data_per_timeline.rs. Log different data on different timelines.

ellipsoids3d_batch 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/ellipsoids3d_batch.rs. Log a batch of Ellipsoids3D.

ellipsoids3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/ellipsoids3d_simple.rs. Log random points and the corresponding covariance ellipsoid.

encoded_image 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/encoded_image.rs. Log a PNG image

entity_path 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/entity_path.rs. Example of different ways of constructing an entity path.

extra_values 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/tutorials/extra_values.rs. Log extra values with a Points2D.

geo_line_strings_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/geo_line_strings_simple.rs. Log a simple geospatial line string.

geo_points_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/geo_points_simple.rs. Log some very simple geospatial point.

graph_directed 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/graph_directed.rs. Log a simple directed graph.

graph_undirected 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/graph_undirected.rs. Log a simple undirected graph.

image_column_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/image_column_updates.rs. Update an image over time, in a single operation.

image_formats 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/image_formats.rs.

image_row_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/image_row_updates.rs. Update an image over time.

image_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/image_simple.rs. Create and log an image

indices 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/concepts/indices.rs. Set different types of indices.

instance_poses3d_combined 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/instance_poses3d_combined.rs. Log a simple 3D box with a regular & instance pose transform.

line_strips2d_batch 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strips2d_batch.rs. Log a batch of 2D line strips.

line_strips2d_segments_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strips2d_segments_simple.rs. Log a couple 2D line segments using 2D line strips.

line_strips2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strips2d_simple.rs. Log a simple line strip.

line_strips2d_ui_radius 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strips2d_ui_radius.rs. Log lines with ui points & scene unit radii.

line_strips3d_batch 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strips3d_batch.rs. Log a batch of 2D line strips.

line_strips3d_segments_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strips3d_segments_simple.rs. Log a simple set of line segments.

line_strips3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strips3d_simple.rs. Log a simple line strip.

line_strips3d_ui_radius 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/line_strips3d_ui_radius.rs. Log lines with ui points & scene unit radii.

mesh3d_indexed 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/mesh3d_indexed.rs. Log a simple colored triangle with indexed drawing.

mesh3d_instancing 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/mesh3d_instancing.rs. Log a simple 3D mesh with several instance pose transforms which instantiate the mesh several times and will not affect its children (known as mesh instancing).

mesh3d_partial_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/mesh3d_partial_updates.rs. Log a simple colored triangle, then update its vertices’ positions each frame.

mesh3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/mesh3d_simple.rs. Log a simple colored triangle.

pinhole_perspective 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/pinhole_perspective.rs. Logs a point cloud and a perspective camera looking at it.

pinhole_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/pinhole_simple.rs. Log a pinhole and a random image.

points2d_random 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points2d_random.rs. Log some random points with color and radii.

points2d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points2d_simple.rs. Log some very simple points.

points2d_ui_radius 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points2d_ui_radius.rs. Log some points with ui points & scene unit radii.

points3d_column_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points3d_column_updates.rs. Update a point cloud over time, in a single operation.

points3d_partial_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points3d_partial_updates.rs. Update specific properties of a point cloud over time.

points3d_random 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points3d_random.rs. Log some random points with color and radii.

points3d_row_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points3d_row_updates.rs. Update a point cloud over time.

points3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points3d_simple.rs. Log some very simple points.

points3d_ui_radius 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/points3d_ui_radius.rs. Log some points with ui points & scene unit radii.

quick_start_connect 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/quick_start/quick_start_connect.rs. Connect to the viewer and log some data.

quick_start_spawn 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/quick_start/quick_start_spawn.rs. Spawn a viewer and log some data.

recording_properties 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/concepts/recording_properties.rs. Sets the recording properties.

scalars_column_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/scalars_column_updates.rs. Update a scalar over time, in a single operation.

scalars_multiple_plots 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/scalars_multiple_plots.rs. Log a scalar over time.

scalars_row_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/scalars_row_updates.rs. Update a scalar over time.

scalars_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/scalars_simple.rs. Log a scalar over time.

segmentation_image_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/segmentation_image_simple.rs. Create and log a segmentation image.

send_recording 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/concepts/send_recording.rs. Send a .rrd to a new recording stream.

series_lines_style 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/series_lines_style.rs. Log a scalar over time.

series_points_style 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/series_points_style.rs. Log a scalar over time.

tensor_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/tensor_simple.rs. Create and log a tensor.

text_document 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/text_document.rs. Log a TextDocument

text_log 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/text_log.rs. Log a TextLog

text_log_integration 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/text_log_integration.rs. Shows integration of Rerun’s TextLog with the native logging interface.

transform3d_axes 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/transform3d_axes.rs. Log different transforms with visualized coordinates axes.

transform3d_column_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/transform3d_column_updates.rs. Update a transform over time, in a single operation.

transform3d_hierarchy 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/transform3d_hierarchy.rs. Log different transforms between three arrows.

transform3d_partial_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/transform3d_partial_updates.rs. Update specific properties of a transform over time.

transform3d_row_updates 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/transform3d_row_updates.rs. Update a transform over time.

transform3d_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/transform3d_simple.rs. Log different transforms between three arrows.

video_auto_frames 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/video_auto_frames.rs. Log a video asset using automatically determined frame references.

video_manual_frames 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/video_manual_frames.rs. Log a video asset using manually created frame references.

view_coordinates_simple 🔒

DO NOT EDIT! This file was autogenerated by docs/snippets/build.rs. The original is in /home/runner/work/rerun/rerun/docs/snippets/all/archetypes/view_coordinates_simple.rs. Change the view coordinates for the scene.

Functions

run